Fixes for PR#192, specifically:
__members__, __dir__(): Return list of members, specifically __doc__,
func_doc, __name__, func_name, func_globals, func_defaults, func_code.

___setattr__(), __delattr__(): throw TypeError if trying to modify a
readonly attribute, and AttributeError if modifying a non-existant
attribute.  In __setattr__(), func_doc and __doc__ are writable.
NOTE: In CPython, func_defaults and func_code are writable, but not in
JPython.

Also, get rid of __class__ and super(__class__) call in ctor.  That
was the old way of initializing this ivar.