generator and checkers vs 22 and 23 of the conversion to new-style classes status: how our own builtin-in types hierarchy matches the corresponding CPython ones.
Some differences are benign: not having __getattribute__ in some places. Some missing things in type.

Other aren't, especially after my dispatch changes, the missing __mod__ are in this category.

__cmp__ vs __eq__ etc difference are an open question.

The problem with the type of __new__ is probably some missing expose_as .

This is still leaving out some important types/descriptors and things that are jython specific
but will need conversion at some point too.

On a related note: indeed the types module is quite wrong at the moment.