First babystep towards replacing the templates with an annotation, @Exposed.
This currently generates two types of classes with asm.  The first comes from
methods with the @Exposed annotation, and is pretty well identical to the
exposed_ classes generated by the templates, though it still lacks all of the
type coercion the templates support.  The second is generated for classes with
@Exposed and it just contains enough information to turn the class into a
Python type.

Next step is to hook this into PyType and turn these into real Python types.