Adding a Jython specific version of CommonErrorNode to use when there are errors in the AST but we want to continue parsing (for now this is only of use by external tools, but I'm considering making it an option since the information can be useful.
Thanks to Allan Davis for the original version of CommonErrorNode.  I've refactored his
version so that PythonErrorNode does not repeat much of CommonErrorNode and put the
creation logic into PythonTreeAdaptor.

I have also added an interface ErrorHandler that allows the error handling in antlr to
be customized.  FailFastHandler implements the default behavior of dying on the first
parse error.  I also made the parse error output more useful, and factored a BaseParser
out of the org/python/antlr/*Parser classes.