Run as many of the 2.7 tests in test_sort as possible
Jython uses the underlying TimSort implementation in
java.util.Arrays.sort (on objects, not primitives), so this means
there are a few Python tests around illegal usages such as random
comparison functions that do not behave the same. Given that this is
for verifying things like not segfaulting, this is not something we
need to duplicate as well.