Expand tab to 4 spaces if a continuation line in the console
Follows the approach suggested in http://bugs.python.org/issue5845

Ideally we would not expand the tab, but insert a tab literal. However
this confuses JLine2 with respect to the cursor position upon
subsequent edits of this line. One can still copy&paste code with tab
literals, although only limited editing is available on the last line
in this scenario.