Improve robustness of socket and SSL support, especially error handling
Note that SSL client connections need to defer the addition of the
inbound handler on the Netty pipeline until after the SSL connection
negotiates the handshake, much as has been done with server SSL. This
will require similar latching to avoid Netty bootstrap issues. In the
meantime, a short sleep (1 ms) seems to be an effective workaround
after doing a handshake. Addresses the issue raised in this email
thread: http://sourceforge.net/p/jython/mailman/message/33541508/ - in
this case, it does not appear to be a question of error handling (Java
exceptions appropriate wrapped as Python socket or SSL errors) as I
suggested in that email.