Better mgmt of lifecycle in ThreadState and PySystemState. Fixes #2321, #2404
ThreadStateMapping tracks the relationship of ThreadState and
java.lang.Thread objects so that ThreadState can be used to augment
threads with additional information needed for Python semantics (a
classic canonical mapping). Indirectly, this augmentation includes
PySystemState, which is exposed to Python code as the sys
module. However if threads are live longer than
ThreadState/PySystemState, as seen in certain app servers for example,
we need to maintain weak relationship to avoid ClassLoader resource
leaks. Previous work to fix #1327 was too aggressive in avoiding this
coupling, which had impact on the public API of the Jython
runtime. Rather than requiring this API to be changed, this commit
dfixes this issue by using a ReferenceQueue to determine when a
PySystemState is no longer referenced outside of a Python call stack,
which means the corresponding ThreadState can be cleaned up.

Based on this PR from Jan Vorwerk
https://bitbucket.org/jython/jython/pull-requests/62/suggested-fix-for-2321/

Also adds a test based on one proposed by Eli Oxman in #2404.