Handle paths used in import and exception text as String or Unicode.
This change set is part of a sequence necessary to handle non-ascii file and
path names correctly. See notes to issue #2356. PyString file paths appearing
during import are handled consistently as UTF-8. Exception processing in Java
allows for non-ascii paths appearing in the traceback, by handling everything
as Java String, although it is not possible to expose this as unicode because
the standard library too often needs bytes.