Fix #2313 test_jython_initializer fails on Windows.
This moves SyspathAppendingInitializer and the service definition for
org.python.core.JythonInitializer into a jar which can be more easily
bundled into the tests. The test is then modified to use the jar instead
of using a class file outside /dist which allows the path to resolve on
Windows.

The jar was built using Java 7 (1.7.0_80).

This also removes the need for a custom compile entry in the build
script.