Remove -E from jython.py (fixes #2707) and improve --print on Linux
Adds quoting and escaping to the output from the launcher when --print
is specified, lost in a previous change, sufficient for
test_jython_launcher to pass. See #2686.

The shebang line of jython.py was not valid on Linux where the entire
line after the first space is treated as one argument, thus
python2.7 is all we can have after usr/bin/env. See #2707.