Treat illegal reflective access in test_gc_jy.
By default we no longer use reflection when traversing cyclic isolates (cyclic
garbage) so that reflective access warnings will not be produced. However,
we enable it, for one class of test case only, on Java versions less than 9.

The clauses of the original test are made individual tests to aid debug.