Fix divergent __str__ and __repr__ of type buffer. Fixes #2719.
This change addresses an issue explored by Ray Ferguson in GH-126 concerning a
lack of clarity, and a potential stack overflow, in the way we expose __str__
and __repr__, and consequent errors in buffer.__str__.

The lack of clarity is dealt with in a note (see PyObject.__repr__). A less
confusing rework of __str__, __repr__ and toString may be in order, but we are
not able to use Ray's approach in GH-126 directly.