Ugrade ANTLR to 3.5.2
The aim is to get rid of the stack dumps that ANTLR 3.1.3 produces
running on Java 8.

This requires some tweaks to the grammar actions since ANTLR 3.5.2
generates List<Object> where it used to have a raw List. Also, the
ANTLR Lexer now returns real EOF tokens that CommonTokenStream
buffers, which we are careful not to consume in PythonTokenSource in
nextToken() or insertImaginaryIndentDedentTokens(). A minimum of
change restores passing behaviour. Some irregularities noticed during
debugging are left for later work.