Refactoring around os.uname() to extract common code.
Use Py.getCommandResult and Py.getenv, the latter so that os.environ is
consulted to pick up changes. Also a slight tweak to the logic allows us to pass
test_uname_win32_ARCHITEW6432 in test_platform, if that test is enabled and
uname_cache disabled (temporarily).

Top-level classes hiding in Hider.java are made nested to placate IDE linters.