Return Java super-class to first position in type.__bases__. Fixes bjo #2445.
The precedence of additions to the proxy types for java.util.List, etc. is
sensitive to the MRO of proxies for Java classes implementing those interfaces.
As initially implemented, this motivated moving the super-class to final
position in type.__bases__, contrary to Java inheritance rles, and this caused
problems in the proxies of classes with complex inheritance. The change here
has us pay greater respect to Java inheritance, having found the placing last
only to be necessary when the (nominal) super-class is java.lang.Object.

This is done with minimal change to the apparatus of proxies and handleMroError,
although close examination during the investigation suggested a number of
improvements peripheral to the fix.