Make README.txt a template for build.xml to fill out.
README.txt is now a template into which version and other information is
written at build time. This change also reorganises the code concerned with
calculating the version string.

A propertry is provided such that:
$ ant "-Dsnapshot.name=TEST" installer
creates an installer for a version identifying as 2.7.2a1-TEST and delivers a
README.txt describing it as a snapshot.