Begin to identify as Jython 2.7.2b2.
As well as changing the version numbers in build scripts as we have
often before, this clarifies (in comments) a numbering policy change by
which developer and snapshot builds identify as the version we are
working *towards*, with a -DEV or -SNAPSHOT suffix, not the last one
released and a "+" suffix. This appears to be the common practice (in
the Maven eco-system).