Add JUnit test for concurrent interpreters.
This is to investigate issue #2834, but is not a fix. It provides a
failing test (except that the failing assertions are commented out).

These assert that importing the same class definition to interpreters
with distinct class loaders results in distinct Python type objects and
Java classes. In fact, the first interpreter ends up sharing them with
all subsequent ones through the common package manager.