Allow source-last-modified != $py.class mtime (addresses #2862)
We allow up to 2 seconds discrepancy in the logic that decides whether
to re-compile .py source to $py.class files. This is to accommodate the
rounding of last-modified time that occurs when storing and extracting
files from a JAR, as during installation. We choose to round times down
when JARring and allow 2 seconds positive difference in the check.

This has benefits beyond #2862, in avoiding spurious recompilation, so
is made a distinct commit.

At the same time the implications of #2862 are arguably not addressed
fully until the JAR cache is made user-local.