Handle IndexError from embedded interpreter
In my environment we are using a Jython Interpreter embedded in a larger application.  After upgrading from a much older version to 2.7.2 we started getting an IndexError that hosed things.  The comment below the modified line notes that sys.argv isn't handled properly in embedded interpreters I suspect that there was a change alsewhere that changed sys.argv from being undefined to being an empty list but that is just speculation.  Regardless this change gets things working in our embedded environment.